+++ lib/src/controls/inputs/slider.dart
@@ -364,9 +364,10 @@ class _CustomTrackShape extends m.RoundedRectSliderTrackShape {
     required Animation<double> enableAnimation,
     required TextDirection textDirection,
     required Offset thumbCenter,
+    Offset? secondaryOffset,
     bool isDiscrete = false,
     bool isEnabled = false,
-    double additionalActiveTrackHeight = 0,
+    double additionalActiveTrackHeight = 2,
   }) {
     return super.paint(
       context,